app/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
openai>=1.40.0
pandas>=2.2.0
openpyxl>=3.1.0
numpy>=1.26.0
python-dotenv>=1.0.0
python-docx>=1.1.0
sse-starlette>=2.0.0
